Create Python Script
This script uses the os.system function to execute a command that restarts the computer. The /r option specifies a restart, /t 0 sets the timeout to 0 seconds, and /f forces running applications to close without warning.

Save this script with a .py extension and run it. Keep in mind that you might need administrative privileges to restart the computer.